requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client
sentence-transformers
openai
pydantic-settings
python-dotenv
